Fixed breaking build after entity renames
