+++ ghost/ghost/src/core/snippets/snippets.repository.bookshelf.ts
@@ -1,5 +1,5 @@
 import {Inject} from '@nestjs/common';
-import {Snippet} from './Snippet';
+import {Snippet} from './snippet.entity';
 import {ISnippetsRepository} from './snippets.repository.interface';
 
 type QueryOptions = {
