+++ ghost/ghost/src/core/snippets/snippets.repository.inmemory.ts
@@ -1,4 +1,4 @@
-import {Snippet} from './Snippet';
+import {Snippet} from './snippet.entity';
 import {ISnippetsRepository} from './snippets.repository.interface';
 
 export class SnippetsRepositoryInMemory implements ISnippetsRepository {
