+++ ghost/ghost/src/core/snippets/snippets.repository.interface.ts
@@ -1,4 +1,4 @@
-import {Snippet} from './Snippet';
+import {Snippet} from './snippet.entity';
 
 export interface ISnippetsRepository {
   findAll(options: {debug?: boolean}): Promise<Snippet[]>;
